Story 1: Modify calculator

+++ JG.FinTechTest/Domain/Services/GiftAidCalculator.cs
@@ -4,7 +4,7 @@ namespace JG.FinTechTest.Domain.Services
 {
     public class GiftAidCalculator : IGiftAidCalculator
     {
-        public int TaxRatePercentage { get; private set; }
+        public decimal TaxRatePercentage { get; private set; }
 
         public GiftAidCalculator(string taxRatePercentage)
         {
@@ -13,7 +13,7 @@ namespace JG.FinTechTest.Domain.Services
 
         public decimal CalculateGiftAid(int donationAmount)
         {
-            return donationAmount * (TaxRatePercentage / (100 - TaxRatePercentage));
+            return donationAmount * (TaxRatePercentage / (100m - TaxRatePercentage));
         }
     }
 }
